job_queue now longer takes ownership of driver II
git-svn-id: http://ert.nr.no/repos/ert@3162 ecb61f28-d21e-0410-a4b6-a828cfe0ca3a

+++ devel/libenkf/src/field_trans.c
@@ -184,11 +184,11 @@ static float trunc_pow10f(float x) {
 
 #define LN_SHIFT 0.0000001
 static float field_trans_ln0( float x ) {
-  return log( x + LN_SHIFT );
+  return logf( x + LN_SHIFT );
 }
 
 static float field_trans_exp0( float x ) {
-  return exp( x ) - LN_SHIFT;
+  return expf( x ) - LN_SHIFT;
 }
 #undef LN_SHIFT
